+++ pkg/pvcresizing/pvc.go
@@ -1,8 +1,6 @@
-package pvresizing
+package pvcresizing
 
 const (
-	Name = "prometheus"
-
 	PrometheusStorageSizeSmall  PrometheusStorageSizeType = "small"
 	PrometheusStorageSizeMedium PrometheusStorageSizeType = "medium"
 	PrometheusStorageSizeLarge  PrometheusStorageSizeType = "large"
@@ -10,15 +8,13 @@ const (
 
 type PrometheusStorageSizeType string
 
-// PrometheusDiskSize returns the desired disk size based on the
-// value of annotation monitoring.giantswarm.io/prometheus-disk-size
-func PrometheusDiskSize(v string) string {
-	value := PrometheusStorageSizeType(v)
+// PrometheusVolumeSize returns the desired disk size based on the
+// param annotationValue: small, medium Or large
+func PrometheusVolumeSize(annotationValue string) string {
+	value := PrometheusStorageSizeType(annotationValue)
 	switch value {
 	case PrometheusStorageSizeSmall:
 		return "30Gi"
-	case PrometheusStorageSizeMedium:
-		return "100Gi"
 	case PrometheusStorageSizeLarge:
 		return "200Gi"
 	default: