+++ service/controller/resource/monitoring/prometheus/resource.go
@@ -16,7 +16,7 @@ import (
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"github.com/giantswarm/prometheus-meta-operator/v2/pkg/pvresizing"
+	"github.com/giantswarm/prometheus-meta-operator/v2/pkg/pvcresizing"
 	"github.com/giantswarm/prometheus-meta-operator/v2/service/controller/resource/generic"
 	"github.com/giantswarm/prometheus-meta-operator/v2/service/key"
 )
@@ -112,9 +112,9 @@ func toPrometheus(ctx context.Context, v interface{}, config Config) (metav1.Obj
 	var walCompression bool = true
 
 	var storage promv1.StorageSpec
-	diskSizeAnnotationValue := cluster.GetAnnotations()[key.PrometheusDiskSizeAnnotation]
-	diskSize := pvresizing.PrometheusDiskSize(diskSizeAnnotationValue)
-	storageSize := resource.MustParse(diskSize)
+	annotationValue := cluster.GetAnnotations()[key.PrometheusDiskSizeAnnotation]
+	volumeSize := pvcresizing.PrometheusVolumeSize(annotationValue)
+	storageSize := resource.MustParse(volumeSize)
 
 	if config.CreatePVC {
 		storage = promv1.StorageSpec{