+++ service/controller/resource/monitoring/pvcresizingresource/delete.go
@@ -1,11 +1,11 @@
-package pvresizingresource
+package pvcresizingresource
 
 import (
 	"context"
 )
 
 func (r *Resource) EnsureDeleted(ctx context.Context, obj interface{}) error {
-	// TODO Check if we need delete
+	// No need to delete anything as PV gets deleted by prometheus-operator
 
 	return nil
 }